Rename function.py to function_type.py. Started cleanup.
