+++ numba/core/types/function_type.py
